refactor: фильтрация фильмов по жанру вынесена в отдельный хук
