+++ src/components/sections/catalog/catalog.jsx
@@ -1,4 +1,4 @@
-import React, {useEffect, useState} from "react";
+import React, {useEffect} from "react";
 import ButtonShowMore from "../../blocks/button-show-more/button-show-more";
 import MoviesList from "../../blocks/movies-list/movies-list";
 import PropTypes from "prop-types";
@@ -9,8 +9,11 @@ import {connect} from "react-redux";
 import {fetchMoviesList} from "../../../store/api-actions";
 import Loader from "../../blocks/loader/loader";
 import {showMoreMovies} from "../../../store/actions";
+import useFilter from "../../hooks/use-filter";
 
 const Catalog = ({movies, isDataLoaded, onLoadData, currentGenre, renderedMoviesCount, onButtonShowMoreClick, currentMovieGenre, filter = false, title = `Catalog`, className}) => {
+
+  const filteredMovies = useFilter(isDataLoaded, movies, currentGenre);
 
   useEffect(() => {
     if (!isDataLoaded) {
@@ -18,14 +21,6 @@ const Catalog = ({movies, isDataLoaded, onLoadData, currentGenre, renderedMovies
     }
   }, [isDataLoaded]);
 
-  const [isFilteredMovies, setFilteredMovies] = useState(movies);
-
-  useEffect(() => {
-    if (isDataLoaded) {
-      const filteredMovies = movies.filter(({genre}) => currentGenre === `All genres` ? genre : genre === currentGenre);
-      setFilteredMovies(filteredMovies);
-    }
-  }, [isDataLoaded, currentGenre]);
   if (!isDataLoaded) {
     return (
       <Loader/>
@@ -33,7 +28,7 @@ const Catalog = ({movies, isDataLoaded, onLoadData, currentGenre, renderedMovies
   }
 
   const similarMovies = movies.filter(({genre}) => genre === currentMovieGenre);
-  const moviesItems = currentMovieGenre ? similarMovies : isFilteredMovies;
+  const moviesItems = currentMovieGenre ? similarMovies : filteredMovies;
 
   const movieList = moviesItems.slice(0, Math.min(moviesItems.length, renderedMoviesCount));
